implement a clickable/expandable fighter list item
A fighter list item component when pressed gets swapped out for a component with more details about the fighter.
Add a "status" property to the fighter object state located at globalState > combatObject > fighters > {fighter}.
The status property is initialized to "n/a" using the emptyFighter template object inside the addFighterView modal.
The status property is currently the only extra info being displayed in the expanded fighter list item comp.
A button is rendered within the expanded fighter list item comp that has access to global state through context.
When the button is pressed, status: "n/a" becomes status: "concentrating" for a simple example.
This button will need to be replaced with a a prettier control: searchable list of icons for quick status updates.